Implemented loading & rendering of level

+++ src/de/earley/pixelengine/entity/Entity.java
@@ -2,6 +2,7 @@ package de.earley.pixelengine.entity;
 
 import de.earley.pixelengine.sprite.Drawable;
 import de.earley.pixelengine.vector.Vector2f;
+import de.earley.pixelengine.vector.Vector2i;
 import de.earley.pixelengine.window.render.Screen;
 
 public abstract class Entity {
@@ -11,7 +12,7 @@ public abstract class Entity {
 	
 	public abstract void update(int delta);
 
-	public abstract void render(Screen screen, Vector2f offset);
+	public abstract void render(Screen screen, Vector2i offset);
 
 	public Vector2f getPosition() {
 		return position.copy();
